requirements.txt
Flask==2.0.2
Flask_HTTPAuth==4.5.0
PyJWT==2.3.0
PyMySQL==1.0.2
python_bcrypt==0.3.2
PyYAML==6.0
redis==4.1.2
